requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dotenv>=1.0.0
rich>=13.7.0
feedparser>=6.0.0
flask>=3.0.0
curl_cffi>=0.7.0
python-jobspy>=1.1.79; python_version < "3.14"
